Improve `NO_STEAMWORKS` support
