+++ map_gui.cpp
@@ -1510,9 +1510,11 @@ void MapGui::updateObjects(CreatureView* view, Renderer& renderer, MapLayout* ma
     const optional<TutorialInfo>& tutorial) {
   selectionSize = view->getSelectionSize();
   playerPosition = view->getPlayerPosition();
+#ifdef USE_STEAMWORKS
   renderer.getSteamInput()->setGameActionLayer(!!playerPosition
       ? MySteamInput::GameActionLayer::TURNED_BASED
       : MySteamInput::GameActionLayer::REAL_TIME);
+#endif
   if (tutorial) {
     tutorialHighlightLow = tutorial->highlightedSquaresLow;
     tutorialHighlightHigh = tutorial->highlightedSquaresHigh;
